requirements.txt
streamlit
pymupdf
groq
faiss-cpu
python-dotenv
numpy
sentence-transformers
